[Add] subscription property분리 및 router생성 및 오류처리
db에서 제대로된 객체를 받아오는지 확인 및 app 및 subscription 분리
