+++ server/src/routes/router.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import auth from './authRouter';
+
+const router = express.Router();
+
+router.use('/auth', auth);
+
+export default router;
